Add integration tests to backend routes

+++ backend/controllers/postController.js
@@ -1,4 +1,5 @@
 const db = require('../models/postModels');
+const userDb = require('../models/userModels');
 const { validationResult, matchedData } = require('express-validator');
 const success = require('../utils/success');
 const failure = require('../utils/failure');
@@ -11,11 +12,11 @@ async function getPost(req, res, next) {
         const post = await db.getPostById(id);
         // if no post is found, return a 404 failure response
         if (!post) {
-            return next(failure(404, 'Failed finding post'));
+            return next(failure(404, 'Post not found'));
         };
 
         // return a 200 success response with the found post
-        return success(res, 200, 'Successfully found post', post);
+        return success(res, 200, 'Post found', post);
     } catch(err) {
         next(err);
     };
@@ -28,7 +29,7 @@ async function getAllPostsForUser(req, res, next) {
     try {
         const posts = await db.getAllPostsForUserById(userId);
         // returns a 200 success response with the found posts
-        return success(res, 200, 'Successfully found posts', posts);
+        return success(res, 200, 'Posts found', posts);
     } catch(err) {
         next(err);
     };
@@ -48,7 +49,7 @@ async function createPost(req, res, next) {
         const { text } = matchedData(req);
         const newPost = await db.createNewPost(text, userId);
         // return a 201 success response with the new post
-        return success(res, 201, 'Successfully created new post', newPost);
+        return success(res, 201, 'Post created', newPost);
     } catch(err) {
         next(err);
     };
@@ -70,7 +71,7 @@ async function updatePost(req, res, next) {
         const post = await db.getPostById(id);
         // if no post is found, return a 404 failure response
         if (!post) {
-            return next(failure(404, 'Failed finding post'));
+            return next(failure(404, 'Post not found'));
         };
 
         // if post author does not match the current user, return a 403 failure response
@@ -81,7 +82,7 @@ async function updatePost(req, res, next) {
         const { text } = matchedData(req);
         const updatedPost = await db.updatePostById(text, id);
         // return a 200 success response with the updated comment
-        return success(res, 200, 'Successfully updated post', updatedPost);
+        return success(res, 200, 'Post updated', updatedPost);
     } catch(err) {
         next(err);
     };
@@ -97,7 +98,7 @@ async function deletePost(req, res, next) {
         const post = await db.getPostById(id);
         // if no post is found, return a 404 failure response
         if (!post) {
-            return next(failure(404, 'Failed finding post'));
+            return next(failure(404, 'Post not found'));
         };
 
         // if post author does not match the current user, return a 403 failure response
@@ -116,10 +117,18 @@ async function deletePost(req, res, next) {
 // obtains all posts created by a peer by peer ID
 async function getPostsForPeer(req, res, next) {
     const id = req.validatedId;
+
     try {
+        // searches user database for peer
+        const peer = await userDb.getUserById(id);
+        // if peer is not found, return a 404 failure response
+        if (!peer) {
+            return next(failure(404, 'User not found'));
+        };
+
         const posts = await db.getPostsForPeerById(id);
         // returns a 200 success response with the found posts
-        return success(res, 200, 'Successfully found posts', posts);
+        return success(res, 200, 'Posts found', posts);
     } catch(err) {
         next(err);
     };
@@ -131,9 +140,16 @@ async function addLikeToPost(req, res, next) {
     const userId = req.user.id;
 
     try {
+        // searches post database for the post by ID
+        const post = await db.getPostById(postId);
+        // if post is not found, return a 404 failure response
+        if (!post) {
+            return next(failure(404, 'Post not found'));
+        };
+
         const newLike = await db.addLikeToPostById(userId, postId);
         // return a 201 success response with the new like
-        return success(res, 201, 'Successfully added like to post', newLike);
+        return success(res, 201, 'Like added to post', newLike);
     } catch(err) {
         next(err);
     };
@@ -145,11 +161,17 @@ async function removeLikeFromPost(req, res, next) {
     const userId = req.user.id;
 
     try {
+        // searches post database for the post by ID
+        const post = await db.getPostById(postId);
+        // if post is not found, return a 404 failure response
+        if (!post) {
+            return next(failure(404, 'Post not found'));
+        };
         // looks through database to ensure the like exists before deleting
         const like = await db.getLikeByIds(userId, postId);
         // if no like is found, return a 404 failure response
         if (!like) {
-            return next(failure(404, 'Failed finding like'));
+            return next(failure(404, 'Like not found'));
         };
 
         const removedLike = await db.removeLikeFromPostById(userId, postId);
